[🚚change]: 게임룸-인게임 라우팅 및 폴더구조 변경, url명 변경 - /gameroom -> /game - GamePage에서 이하 컴포넌트들을 렌더링 합니다
