+++ src/pages/GamePage/GameWaitingRoom/GameModeInfo.tsx
