+++ src/pages/GamePage/GameWaitingRoom/GameRoomSetting.tsx
